requirements.txt
# django apps
django
drf-yasg
djangorestframework
djangorestframework-simplejwt
django-jazzmin
django-ckeditor-5
django-debug-toolbar
django-configurations
django-cors-headers
whitenoise
gunicorn

# Third Party apps
VerbalExpressions
python-dotenv
str2bool
pycountry
phonenumbers
qrcode[pil]
Pillow
seavoyage
pydantic


# Database
dj-database-url
psycopg2-binary

# Test
pytest
pytest-django
pytest-cov
coverage
coverage-badge
